Move Donation to a view modifier
